TST/REF: misplaced Index.putmask tests
